risk-orchestrator/pyproject.toml
[tool.poetry]
name = "risk-orchestrator"
version = "0.1.0"
description = "Centralized risk & order routing for autonomous trading fleet"
authors = ["AIAster Team"]

[[tool.poetry.packages]]
include = "risk_orchestrator"
from = "src"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
httpx = "^0.27.0"
redis = {extras = ["hiredis"], version = "^5.0.0"}
google-cloud-pubsub = ">=2.13.0"
pydantic-settings = "^2.4.0"
python-dotenv = "^1.0.0"
loguru = "^0.7.0"
prometheus-fastapi-instrumentator = "^6.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
